price again

+++ market14B/src/main/java/org/market/DomainLayer/backend/UserPackage/User.java
@@ -58,7 +58,7 @@ public abstract class User {
         return "added to cart";
     }
 
-    public String addToCartOffer(Integer product, int storeId, int price) throws Exception {
+    public String addToCartOffer(Integer product, int storeId, double price) throws Exception {
         shoppingCart.addToCartOffer(this.username, product, price, storeId);
         UserController.LOGGER.info("added to cart");
         return "added to cart";
